feat(num): float prints with decimal, JSON subtype fidelity, int()/float() (§4)
- Float always renders a decimal: print(5.0)->5.0, [1.0,2.5], -0.0; int still 5;
  inf/nan unchanged; non-integral floats print exactly as before. One shared
  value::format_float, so str/print/templates/both engines agree.
- JSON: stringify(float)->5.0, stringify(int)->5; parse splits int/float by syntax
  (parse∘stringify preserves subtype).
- int()/float() conversion builtins: float->int truncates toward zero; string->Tier-1
  [v,err]; wrong type->clean Tier-2 panic. FIXED reviewer-found bug: int(2^63) was
  silently saturating to i64::MAX (i64::MAX as f64 rounds up to 2^63) -> strict upper
  bound -(i64::MIN as f64); regression test added.
- Lexer: InvalidDigit vs OutOfRange messages (0o19 -> 'invalid digit').
- Gate-7 corpus migration: 9 goldens + inline expectations to float-with-decimal;
  several integer-valued stdlib returns switched Float->Int.

Independent reviewer: PASS after the boundary fix. VERIFIED: build+clippy clean both
configs; vm_differential 357/0; full suite green both configs; Gate 5 zero type-*.

NUM float-printing+conversions. Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>

+++ src/stdlib/ai/response.rs
@@ -136,7 +136,8 @@ fn usage_object(n: &NeutralResponse) -> Value {
 
 fn opt_num(v: Option<i64>) -> Value {
     match v {
-        Some(n) => Value::Float(n as f64),
+        // NUM §4: a token count is an `int`.
+        Some(n) => Value::Int(n),
         None => Value::Nil,
     }
 }
@@ -167,7 +168,8 @@ pub(crate) fn embed_object(resp: &genai::embed::EmbedResponse, many: bool) -> Va
     usage.insert(
         "inputTokens".to_string(),
         match resp.usage.prompt_tokens {
-            Some(n) => Value::Float(n as f64),
+            // NUM §4: a token count is an `int`.
+            Some(n) => Value::Int(n as i64),
             None => Value::Nil,
         },
     );
